% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_spline_functions.R
\name{ana_B}
\alias{ana_B}
\title{ana_B}
\usage{
ana_B(B, ext_knots, int_knots, order, knots, num_basis)
}
\arguments{
\item{B}{matrix with value of basic splines (rows) for given set of time points (columns)}

\item{ext_knots}{set of extended knots}

\item{int_knots}{set of time points, defined such that there are degree -1 points between
each consecutive knots}

\item{order}{degree of the basic spline polynomial + 1}

\item{knots}{set of knots}

\item{num_basis}{number of basic splines (number of knots + degree - 1)}
}
\value{
matrix b.hat with for each b_spline the coefficients of the polynomial
for one combination of knots
matrix support_bs which indicates on which knots a basic spline is defined
}
\description{
From a matrix B, containing the values of each basic spline (rows) for the knots
and (degree - 1) points between the knots, this function calculates the support
of each basic spline and the coefficients of the polynomial of each basic spline
between each set of knots.
}
