% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_spline_functions.R
\name{analytical_bspline}
\alias{analytical_bspline}
\title{analytical_bspline}
\usage{
analytical_bspline(b_hat, a, b_support, t_new, knots)
}
\arguments{
\item{b_hat}{matrix with coefficients of the b_spline for each interval between knots}

\item{a}{coefficients of the b spline}

\item{b_support}{matrix indicationg for which interval between knots a b spline is defined}

\item{t_new}{the value for which the b spline is to be calculated}

\item{knots}{set of knots}
}
\value{
spline value at t_new
}
\description{
function to calculate full spline for new values, given the coefficients of the
polynomials that define the spline (b_hat) the coefficients of the spline (a)
and the support of each polynomial (b_support)
}
