% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{bernoulli}
\alias{bernoulli}
\title{Example function to run the example Bernoulli model}
\usage{
bernoulli(y)
}
\arguments{
\item{y}{The binary outcome of the trials.}
}
\value{
The draws for the Bernoulli chance-of-success `theta` parameter.
}
\description{
Example function to run the example Bernoulli model
}
