% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_spline_functions.R
\name{build_b_spline}
\alias{build_b_spline}
\title{build_b_spline}
\usage{
build_b_spline(t, ext_knots, ind, order)
}
\arguments{
\item{t}{time points at which to calculate the spline values}

\item{ext_knots}{extended series of knots}

\item{ind}{index over the number of basic splines}

\item{order}{order of the spline = degree + 1}
}
\value{
a vector of the value of the b spline of ind at times t
}
\description{
Construct basic b spline of (ind) that indicates the values at time t for this
spline. Used from Milad Kharratzadeh (https://mc-stan.org/users/documentation/case-studies/splines_in_stan.html))
}
