% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contact_all}
\alias{contact_all}
\title{Synthetic contact matrices per country by Prem et al. (2021)}
\format{
\subsection{\code{contact_all}}{

A list of 177 arrays
\describe{
\item{country}{synthetic contact matrix per country}
}
}
}
\source{
\url{https://doi.org/10.1371/journal.pcbi.1009098}
}
\usage{
contact_all
}
\description{
Synthetic contact matrices per country by Prem et al. (2021)
}
\keyword{datasets}
