% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_spline_functions.R
\name{create_B}
\alias{create_B}
\title{Create_B}
\usage{
create_B(t, order, knots)
}
\arguments{
\item{t}{time points at which to calculate the b spline}

\item{order}{order of the spline = degree + 1}

\item{knots}{set of knots}
}
\value{
matrix B with each row representing a basic spline and each column a value of t
}
\description{
Function to run the algorithm (inspired from Milad Kharratzadeh (https://mc-stan.org/users/documentation/case-studies/splines_in_stan.html))
and create matrix B to define a basic spline
}
