% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_simulating_functions.R
\name{create_contactmatrix_GE}
\alias{create_contactmatrix_GE}
\title{create_contactmatrix_GE}
\usage{
create_contactmatrix_GE(contact_all)
}
\arguments{
\item{contact_all}{contact data to calculate the contact matrix using Prem et al (2021)}
}
\value{
List with contact matrix (1) and the simulated population (2)
}
\description{
Function to construct a contact matrix based on Prem et al. (2021) for Switzerland and the age
distribution of the Canton of Geneva (2021). Age groups are fixed at 0-19, 20-59
and 60+.
}
