% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_sampling_functions.R
\name{create_function_initial_values}
\alias{create_function_initial_values}
\title{create_function_initial_values}
\usage{
create_function_initial_values(
  data_list,
  homogeneous = TRUE,
  type = "spline",
  sampling = "qp"
)
}
\arguments{
\item{data_list}{(list) List of variables needed to run the Stan model as generated by
"standata".}

\item{homogeneous}{(logical) TRUE if an unstratified model should be fitted and FALSE if
the fitted model should be stratified.}

\item{type}{(string) Type of time-dependence that is included: "BM" (Brownian motion),
"spline" (splines) or "GP" (Gaussian process).}

\item{sampling}{(string) The sampling distribution that should be used: "p" (Poisson),
"qp" (quasi-Poisson) or "negbin" (negative binomial).}
}
\value{
List of initial values for running the specified Stan model.
}
\description{
create_function_initial_values
}
