% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GE_data}
\alias{GE_data}
\title{SARS-CoV-2 related data for the Canton of Geneva, Switzerland
in 2020.}
\format{
\subsection{\code{GE_data}}{

List of 5 items
\describe{
\item{seroprevalence study 1}{number of tests performed, number of positive tests per age-group}
\item{seroprevalence study 2}{number of tests performed, number of positive tests per age-group}
\item{seroprevalence study 3}{number of tests performed, number of positive tests per age-group}
\item{weekly stratified cases}{Number of positive PCR tests per week per age-class in Geneva during 2020}
\item{population distribution Geneva}{Number of inhabitants per age-class}
}
}
}
\source{
\url{https://doi.org/10.1371/journal.pcbi.1009098}
}
\usage{
GE_data
}
\description{
SARS-CoV-2 related data for the Canton of Geneva, Switzerland
in 2020.
}
\keyword{datasets}
