% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_simulating_functions.R
\name{get_incidence}
\alias{get_incidence}
\title{get_incidence}
\usage{
get_incidence(out, popsize, p_detect1, p_detect2, t_detectSwitch)
}
\arguments{
\item{out}{output of SEIR-ode model}

\item{popsize}{population size}

\item{p_detect1}{ascertainment rate of first time-period}

\item{p_detect2}{ascertainment rate of second time-period}

\item{t_detectSwitch}{time at which ascertainment-rate is switched}
}
\value{
vector with incidence per week
}
\description{
Function to calculate the ascertained COVID-19 incidence of simulated data using the output
of SEIR-ode function. Incidence is defined as those moving from compartment
I to R.
}
