% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_simulating_functions.R
\name{get_incidence_strat}
\alias{get_incidence_strat}
\title{get_incidence_strat}
\usage{
get_incidence_strat(R, popsize, p_detect1, p_detect2, t_detectSwitch)
}
\arguments{
\item{R}{array with number of individuals per age-class per week}

\item{popsize}{vector of population size per age-class}

\item{p_detect1}{ascertainment rate of first time-period}

\item{p_detect2}{ascertainment rate of second time-period}

\item{t_detectSwitch}{time at which ascertainment-rate is switched}
}
\value{
array with incidence per age-class per week
}
\description{
Function to calculate the ascertained COVID-19 incidence of stratified simulated
data using the output of SEIR-ode function. Incidence is defined as those moving from compartment
I to R.
}
