% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEZEIK-package.R
\docType{package}
\name{HETTMO-package}
\alias{HETTMO}
\alias{HETTMO-package}
\title{HETTMO: HEterogeneous Time-dependent Transmission MOdel (HETTMO)}
\description{
This package supplies functions that can efficiently implement time-dependent transmission in a stratified SEIR model in Stan. It includes three different methods of implementing the time dependence: Brownian motion, splines and a Gausian process. More details on the model can be found in the publication linked to this R-package (LINK).
}
\author{
\strong{Maintainer}: Judith Bouman \email{judith.bouman@unibe.ch} (\href{https://orcid.org/0000-0003-3781-0675}{ORCID})

Authors:
\itemize{
  \item Julien Riou (\href{https://orcid.org/0000-0002-3808-2852}{ORCID})
  \item Anthony Hauser (\href{https://orcid.org/0000-0002-7221-1929}{ORCID})
}

}
\keyword{internal}
