% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_simulating_functions.R
\name{ind}
\alias{ind}
\title{Function to determine the number of the equation corresponding to given age class}
\usage{
ind(comp, age_sex, num_class = 3)
}
\arguments{
\item{comp}{required compartment of SEIR model (1=S, 2=E, 3=I, 4 = R)}

\item{age_sex}{required age-classe (1= 0-19, 2=20-59, 3=60+)}

\item{num_class}{total number of modelled age-classes}
}
\value{
number of the ODE function corresponding to "comp" and "age_sex"
}
\description{
Function to determine the number of the equation corresponding to given age class
}
