% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_simulating_functions.R
\name{ODE_simple}
\alias{ODE_simple}
\title{Function with system of ODE equations for SEIR model}
\usage{
ODE_simple(time, y, params)
}
\arguments{
\item{time}{time in weeks}

\item{y}{number of individuals in each compartment at time}

\item{params}{set of parameters for SEIR model}
}
\value{
list of number of individuals per compartment
}
\description{
Function with system of ODE equations for SEIR model
}
