% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_simulating_functions.R
\name{ODE_spline_strat}
\alias{ODE_spline_strat}
\title{Function with system of ODE equations for age-stratified SEIR model with
spline-based time-dependent transmission rate.}
\usage{
ODE_spline_strat(time, y, params)
}
\arguments{
\item{time}{time in weeks}

\item{y}{number of individuals in each compartment at time}

\item{params}{set of parameters as generated by "set_parameters"}
}
\value{
list of number of individuals per compartment
}
\description{
Function with system of ODE equations for age-stratified SEIR model with
spline-based time-dependent transmission rate.
}
