% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_plotting_functions.R
\name{plot_compare}
\alias{plot_compare}
\title{plot_compare}
\usage{
plot_compare(fitBM, fitSpline, fitGP, data, params, transmission_prob_sim)
}
\arguments{
\item{transmission_prob_sim}{}
}
\description{
plot_compare
}
