% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/021_new_plotting_functions.R
\name{plot_compare_sampling_GE}
\alias{plot_compare_sampling_GE}
\title{Title}
\usage{
plot_compare_sampling_GE(fit1, fit2, fit3, data, GE_data)
}
\arguments{
\item{GE_data}{}
}
\description{
Title
}
