% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_plotting_functions.R
\name{plot_comparesim}
\alias{plot_comparesim}
\title{plot_comparesim}
\usage{
plot_comparesim(fit1, fit2, fit3, data)
}
\arguments{
\item{fit1}{modelfit 1 of 3, result of "sampling_model"}

\item{fit2}{modelfit 2 of 3, result of "sampling_model"}

\item{fit3}{modelfit 3 of 3, result of "sampling_model"}

\item{data}{data to which model was fit, vector with value per week}
}
\value{
plot
}
\description{
Function that produces Figure to visualize and compare three different model fits.
}
