% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_plotting_functions.R
\name{plot_fitsim}
\alias{plot_fitsim}
\title{plot_fitsim}
\usage{
plot_fitsim(fit, data, params)
}
\arguments{
\item{fit}{modelfit, result of "sampling_model"}

\item{data}{data to which model was fit, vector with value per week}

\item{params}{list of parameters used to fit model}
}
\description{
Function that produces Figure to visualize model fit.
}
