% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_plotting_functions.R
\name{plot_fitsim_strat}
\alias{plot_fitsim_strat}
\title{plot_fitsim_strat}
\usage{
plot_fitsim_strat(fit, sim_data, params = params, transmission_prob_sim)
}
\arguments{
\item{fit}{fit of stratified model, result of "sampling_model"}

\item{sim_data}{array of simulated positive PCR cases per week and age-class.}

\item{params}{list of parameters used to simulate the data, result of "set_parameters".}

\item{transmission_prob_sim}{array of simulated rho value per week per age-class.}
}
\value{
Figure of result of simulated, stratified model.
}
\description{
Function that produces Figure to visualize stratified model fit to simulated data.
}
