% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/021_new_plotting_functions.R
\name{plot_unstrat_model_fit}
\alias{plot_unstrat_model_fit}
\title{Title}
\usage{
plot_unstrat_model_fit(fit1, fit2, fit3, data_cases, data_prob, params)
}
\arguments{
\item{params}{}
}
\description{
Title
}
