% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_spline_functions.R
\name{probability_transmission_spline}
\alias{probability_transmission_spline}
\title{probability_transmission_spline}
\usage{
probability_transmission_spline(t, order, knots, a, beta)
}
\arguments{
\item{t}{time}

\item{order}{degree of the basic spline polynomial + 1}

\item{knots}{set of knots}

\item{a}{coefficients for the spline}

\item{beta}{intercept of the spline function}
}
\description{
Function that produces the probability of transmission for times t for a spline with
given degree, knots and coefficients a
}
