% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_saving_loading.R
\name{read_stan_result}
\alias{read_stan_result}
\title{Read in model results}
\usage{
read_stan_result(base_name, dir)
}
\arguments{
\item{dir}{data directory where the object is saved}
}
\value{
list of posterior samples, prior samples, model parameters and diagnostics
}
\description{
Read in model results
}
