% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_simulating_functions.R
\name{reciprocal}
\alias{reciprocal}
\title{reciprocal}
\usage{
reciprocal(m, N)
}
\arguments{
\item{m}{original contact matrix}

\item{N}{vector with number of individuals per age-class}
}
\value{
rebalanced contact matrix
}
\description{
Function to rebalance contactmatrix based on the number of individuals in each
class
}
