% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_simulating_functions.R
\name{rqpois}
\alias{rqpois}
\title{rqpois}
\usage{
rqpois(n, mu, theta)
}
\arguments{
\item{n}{number of samples required}

\item{mu}{mean of distribution sampling from}

\item{theta}{overdispersion parameter of quasi-poisson distribution}
}
\description{
Function to randomly sample from a quasi-poisson distribution
}
