% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_sampling_functions.R
\name{sampling_model}
\alias{sampling_model}
\title{sampling_model}
\usage{
sampling_model(
  data_list,
  model = "qpoisson_spline_unstrat.stan",
  save_result,
  type = "spline",
  sampling = "qp",
  homogeneous = TRUE,
  sampler = 4,
  rtol = 1e-06,
  atol = 1e-06,
  max_num_steps = 1000
)
}
\arguments{
\item{data_list}{(list) List of variables for the selected Stan model as generated by "standata".}

\item{model}{(string) The selected model for fitting the data: default option is "qpoisson_spline_unstrat.stan".
Other options: "poisson_spline_unstrat", "negbin_spline_unstrat", "qpoisson_BM_unstrat",
"qpoisson_GP_unstrat", "qpoisson_spline_strat" and "qpoisson_spline_strat_GE".}

\item{save_result}{(string) The path to the file where the result of the model fit should be saved}

\item{type}{(string) type of time-dependence that is included: "BM" (Brownian motion),
"spline" (splines) or "GP" (Gaussian process).}

\item{sampling}{(string) The sampling distribution that should be used: "p" (Poisson),
"qp" (quasi-Poisson) or "negbin" (negative binomial).}

\item{homogeneous}{(logical) TRUE if an unstratified model should be fitted and FALSE if
the fitted model should be stratified.}

\item{sampler}{(integer) Choice of sampler used to solve the ODE-system within the Stan model,
options are: (0) rk45, (1) adams, (2) bdf, (3) ckrk and (4) Trapeziodal solver.}

\item{rtol}{(real) The relative tolerance for solvers 0-3.}

\item{atol}{(real) The absoluted tolerance for solvers 0-3.}

\item{max_num_steps}{(positive interger) Maximum number of steps in mcmc.}
}
\value{
(list) List with prior samples (samples_prior), posterior samples (samples_posterior),
data list used for analysis (data_list) and diagnostics of the model fit (diagnostics).
}
\description{
Function that fits the selected stan model to the supplied case data.
}
