% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_saving_loading.R
\name{save_stan_result}
\alias{save_stan_result}
\title{Save our model results without cmdstan dependency}
\usage{
save_stan_result(stan_result, base_name, dir)
}
\arguments{
\item{stan_result}{model result that needs to be saved}

\item{base_name}{base of naming the result files}

\item{dir}{directory where files should be saved}
}
\description{
Save our model results without cmdstan dependency
}
