% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_simulating_functions.R
\name{set_parameters}
\alias{set_parameters}
\title{set_parameters}
\usage{
set_parameters(stratified = FALSE, contact_all)
}
\arguments{
\item{stratified}{should the simulated data be stratified by age?}

\item{contact_all}{contact data to calculate the contact matrix using Prem et al (2021)}
}
\value{
list of parameters
}
\description{
Function to create set of parameters to simulate data as presented in the
publication.
}
