% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_simulating_functions.R
\name{simulate_data}
\alias{simulate_data}
\title{simulate_data}
\usage{
simulate_data(params, time = 2, ts)
}
\arguments{
\item{params}{list of parameters as generated by "set_parameters"}

\item{time}{type of time-dependence in the simulation (2=spline) (currently only one option)}

\item{ts}{vector with time points at which simulated data is required}
}
\value{
list with following elements: (1) vector with number of simulated positive PCR-tests
(2) vector with effective reproduction number over time
(3) vector with rho(t) over time
(4) vector with number of individuals in compartment I
(5) vector with number of individuals in compartment E
(6) vector with number of individuals in compartment R
(7) number of test in the seroprevalence study that were positive
}
\description{
Function to simulate unstratified data as presented in the publication
}
