% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_simulating_functions.R
\name{simulate_strat_data}
\alias{simulate_strat_data}
\title{simulate_strat_data}
\usage{
simulate_strat_data(params, time = 2, ts)
}
\arguments{
\item{params}{list of parameters as generated by "set_parameters"}

\item{time}{type of time-dependence in the simulation (2=spline) (currently only one option)}

\item{ts}{vector with time points at which simulated data is required}
}
\value{
list with following elements: (1) array with number of simulated positive PCR-tests per age-class
(2) array with effective reproduction number over time per age-class
(3) number of test in the seroprevalence study that were positive per age-class
(4) array with ascertained incidence per week per age-class
(5) array with output of ODE model for SEIR per age-class
}
\description{
Function to simulate stratified data as presented in the publication
}
