% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_nonstratified}
\alias{simulated_nonstratified}
\title{Simulated non-stratified data, using the parameters as specified in the manuscript (function "set_parameters")}
\format{
\subsection{\code{simulated_nonstratified}}{

List of 4 items
\describe{
\item{sim_data}{simulated positive PCR tests per week}
\item{re_sim}{Effective reproduction number per week}
\item{n_infected_survey}{number of positive serological tests}
\item{prob_trans_sim}{Rho(t) per week}
}
}
}
\usage{
simulated_nonstratified
}
\description{
Simulated non-stratified data, using the parameters as specified in the manuscript (function "set_parameters")
}
\keyword{datasets}
