% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_stratified}
\alias{simulated_stratified}
\title{Simulated stratified data, using the parameters as specified in the manuscript (function "set_parameters")}
\format{
\subsection{\code{simulated_stratified}}{

List of 3 items
\describe{
\item{sim_data}{array simulated positive PCR tests per week and age-class}
\item{beta_sim}{Rho(t) per week and age-class}
\item{n_infected_survey}{number of positive serological tests per age-class}
}
}
}
\usage{
simulated_stratified
}
\description{
Simulated stratified data, using the parameters as specified in the manuscript (function "set_parameters")
}
\keyword{datasets}
