% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_sampling_functions.R
\name{standata}
\alias{standata}
\title{standata}
\usage{
standata(
  data,
  homogeneous = TRUE,
  simulated = TRUE,
  type = "spline",
  sampling = "qp",
  seroprev_dat,
  additional_data
)
}
\arguments{
\item{data}{(vector, array) Vector (unstratified model) or array (stratified model) with the
(weekly) case data that the model should be fitted to.}

\item{homogeneous}{TRUE if an unstratified model should be fitted and FALSE if
the fitted model should be stratified}

\item{simulated}{TRUE if simulated data is used and FALSE if}

\item{type}{type of time-dependence that is included: "BM" (Brownian motion),
"spline" (splines) or "GP" (Gaussian process)}

\item{sampling}{the sampling distribution that should be used: "p" (Poisson),
"qp" (quasi-Poisson) or "negbin" (negative binomial)}

\item{seroprev_dat}{dataframe of seroprevalence data if homogeneous model is used,
should include one column called "num_tested" with the number of tests performed and
one column called "num_pos_tests" with the number of positive tests}

\item{additional_data}{a list with additional data elements, only relevant when
the stratified model is used. For the simulated data, additional
data only contains the data from Prem et al. (2021) to create the contact matrix.
For the real case data it needs to include all seroprevalence datasets and a
vector called "pop_GE" with the population distribution over the age-classes for the
Canton of Geneva.}
}
\value{
List with all input variables needed to run the Stan model
}
\description{
Function to generate a list with all input variables needed to run the Stan
model as specified by the input variables
}
