% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_sampling_functions.R
\name{create_function_initial_values}
\alias{create_function_initial_values}
\title{create_function_initial_values}
\usage{
create_function_initial_values(
  data_list,
  homogeneous = TRUE,
  type = "spline",
  sampling = "qp"
)
}
\arguments{
\item{sampling}{}
}
\description{
create_function_initial_values
}
