% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_plotting_functions.R
\name{plot_fitsim_strat_realdata}
\alias{plot_fitsim_strat_realdata}
\title{plot_fitsim_strat_realdata}
\usage{
plot_fitsim_strat_realdata(fit, data, cum_inc, seroprev)
}
\arguments{
\item{fit}{modelfit of stratified model, result of "sampling_model"}

\item{data}{data to which model was fit, array with value per week and age-class}

\item{cum_inc}{vector of cumulative incidence over time}

\item{seroprev}{seroprevalence data used for model fit}
}
\value{
Figure
}
\description{
Function that produces Figure to visualize stratified model fit to data from Geneva.
}
