% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_sampling_functions.R
\name{sampling_model}
\alias{sampling_model}
\title{sampling_model}
\usage{
sampling_model(
  data_list,
  model = "poisson_spline_unstrat.stan",
  save_result,
  type = "spline",
  sampling = "qp",
  homogeneous = TRUE,
  sampler = 4,
  rtol = 1e-06,
  atol = 1e-06,
  max_num_steps = 1000
)
}
\arguments{
\item{...}{}
}
\description{
sampling_model
}
