% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_sampling_functions.R
\name{standata}
\alias{standata}
\title{standata}
\usage{
standata(
  data,
  homogeneous = TRUE,
  simulated = TRUE,
  type = "spline",
  sampling = "qp",
  seroprev_dat,
  additional_data
)
}
\arguments{
\item{additional_data}{}
}
\description{
standata
}
